updated merge yearly

+++ clearskypy/downloader/download.py
@@ -197,6 +197,7 @@ class SocketManager:
         auth: dict = None,
         output_directory: Union[str, Path] = None,
         thread_num: int = 5,
+        merge_yearly: bool = True,
     ):
         """
         MERRA2 universal download.
@@ -214,7 +215,7 @@ class SocketManager:
         auth : dict,
         output_directory : Union[str, Path]
         thread_num : int
-
+        merge_yearly: bool
         """
         if not isinstance(output_directory, Path):
             log_file = Path(output_directory)
@@ -227,12 +228,12 @@ class SocketManager:
 
         dates = []
         for date in self.iter_days(datetime.date(initial_year, initial_month, initial_day), datetime.date(final_year, final_month, final_day)):
-            if self.build_remote_filename(merra2_collection, date, params) in log:
+            if self.build_remote_filename(merra2_collection, date, params) + str(merge_yearly) in log:
                 logging.info("Skipping existing file " + self.build_remote_filename(merra2_collection, date, params) + " from " + merra2_collection["esdt_dir"])
                 continue
             else:
                 logging.info("Preparing new file " + self.build_remote_filename(merra2_collection, date, params) + " from " + merra2_collection["esdt_dir"])
-                log.append(self.build_remote_filename(merra2_collection, date, params))
+                log.append(self.build_remote_filename(merra2_collection, date, params) + str(merge_yearly))
             dates.append(date)
 
         if len(dates) > 0:
@@ -373,6 +374,8 @@ class SocketManager:
             filepath, filename = os.path.split(output_file)
             merging_file = os.path.join(path_data, filename)
             nc_files.append(merging_file)
+        else:
+            merging_file = None
         nc_files.sort()
         logging.info("- Find Files [{0}] To Merge...".format(','.join([os.path.split(name)[1] for name in nc_files])))
 
@@ -563,7 +566,7 @@ class SocketManager:
         np.save(log_file, np.array(log))
 
         logging.info("- Removing Old Merged Files...")
-        if os.path.exists(merging_file):
+        if merging_file is not None and os.path.exists(merging_file):
             os.remove(merging_file)
         logging.info("---- Finish Merging In Time ----")
 
@@ -585,6 +588,7 @@ class SocketManager:
         output_dir: Union[str, Path] = None,
         auth: dict = None,
         delete_temp_dir: bool = True,
+        merge_yearly: bool = True,
         thread_num: Optional[int] = 5,
     ):
         """MERRA2 daily download and conversion.
@@ -634,6 +638,7 @@ class SocketManager:
             Dictionary contains login information.
             {"uid": "USERNAME", "password": "PASSWORD"}
         delete_temp_dir : bool
+        merge_yearly: bool
         thread_num : Optional[int]
             Number of Files to be downloaded simutanously.
 
@@ -717,26 +722,28 @@ class SocketManager:
                     auth=auth,
                     params=parameter,
                     thread_num=thread_num,
+                    merge_yearly=merge_yearly,
                 )
-                # Name the output file
-                if initial_year == final_year:
-                    file_name_str = "{0}_{1}_merra2_reanalysis_{2}.nc"
-                    out_file_name = file_name_str.format(collection_name, merra2_var_dict["esdt_dir"], str(initial_year))
-                else:
-                    file_name_str = "{0}_{1}_merra2_reanalysis_{2}-{3}.nc"
-                    out_file_name = file_name_str.format(
-                        collection_name, merra2_var_dict["esdt_dir"], str(initial_year), str(final_year)
+                if merge_yearly:
+                    # Name the output file
+                    if initial_year == final_year:
+                        file_name_str = "{0}_{1}_merra2_reanalysis_{2}.nc"
+                        out_file_name = file_name_str.format(collection_name, merra2_var_dict["esdt_dir"], str(initial_year))
+                    else:
+                        file_name_str = "{0}_{1}_merra2_reanalysis_{2}-{3}.nc"
+                        out_file_name = file_name_str.format(
+                            collection_name, merra2_var_dict["esdt_dir"], str(initial_year), str(final_year)
+                        )
+                    out_file = Path(output_dir).joinpath(out_file_name)
+                    # Extract variable
+                    self.daily_netcdf(
+                        temp_dir_download,
+                        out_file,
+                        collection_name,
+                        initial_year,
+                        final_year,
+                        merra2_var_dict
                     )
-                out_file = Path(output_dir).joinpath(out_file_name)
-                # Extract variable
-                self.daily_netcdf(
-                    temp_dir_download,
-                    out_file,
-                    collection_name,
-                    initial_year,
-                    final_year,
-                    merra2_var_dict
-                )
             if delete_temp_dir:
                 shutil.rmtree(temp_dir_download)
             else:
